CodeDeck/CodeDeckService/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
llama-cpp-python>=0.2.11
pydantic>=2.4.2
jinja2>=3.1.2
python-multipart>=0.0.6
psutil>=5.9.0
piper-tts>=1.2.0
aiohttp>=3.8.5 